Add more validations, exceptions

+++ PhDSystem.Data/Repositories/ExamRepository.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using PhDSystem.Data.Entities;
+using PhDSystem.Data.Exceptions;
 using PhDSystem.Data.Models.Exams;
 using PhDSystem.Data.Repositories.Interfaces;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -14,7 +16,7 @@ namespace PhDSystem.Data.Repositories
 
         public ExamRepository(PhdSystemDbContext context)
         {
-            _context = context;
+            _context = context ?? throw new ArgumentNullException(nameof(context));
         }
 
         public async Task AddExam(Exam exam)
@@ -25,14 +27,20 @@ namespace PhDSystem.Data.Repositories
 
         public async Task DeleteExam(int examId)
         {
-            var exam = await _context.Exams.Where(e => e.Id.Equals(examId)).SingleOrDefaultAsync();
+            var exam = await _context.Exams.SingleOrDefaultAsync(e => e.Id.Equals(examId));
+
+            if(exam == null)
+            {
+                throw new NotFoundException(typeof(Exam).Name, examId);
+            }
+
             _context.Exams.Remove(exam);
             await _context.SaveChangesAsync();
         }
 
         public async Task<IEnumerable<ExamDetails>> GetExams(int studentId)
         {
-            var exams = await _context.Exams.Where(e => e.StudentId == studentId).Select(x => new ExamDetails
+            return await _context.Exams.Where(e => e.StudentId == studentId).Select(x => new ExamDetails
             {
                 Id = x.Id,
                 StudentId = x.StudentId,
@@ -41,8 +49,6 @@ namespace PhDSystem.Data.Repositories
                 Date = x.Date,
                 Grade = x.Grade
             }).ToListAsync();
-
-            return exams;
         }
     }
 }
